refactor: co-locate search cache lifecycle with its store state
The filtering logic was split across three modules: taskSearch (pure text
search), taskFiltering (a grab-bag of cache lifecycle + board-access +
complexity helpers), and taskStore.filters (which owned the searchCache
state but reached into taskFiltering and passed that state back in). To
apply filters you had to reason across all three.

Fold taskFiltering's helpers into taskStore.filters, beside the searchCache
state and the applyFilters action that drive them, and delete taskFiltering.
taskSearch stays a pure, reusable text-search module. Also rename the free
validateBoardAccess(tasks) to filterAccessibleTasks to end its name clash
with the store action validateBoardAccess(boardId).

Pure relocation — no behaviour change; the existing cache tests move with
the code and stay green.
