+++ src/lib/stores/taskStore.filters.ts
@@ -8,15 +8,6 @@ import { taskDB } from '@/lib/utils/database';
 import { useSettingsStore } from '@/lib/stores/settingsStore';
 import { sanitizeSearchQuery, searchRateLimiter } from '@/lib/utils/security';
 import { searchTasks } from '@/lib/utils/taskSearch';
-import {
-  validateBoardAccess,
-  generateCacheKey,
-  checkCache,
-  cacheResults,
-  cleanupExpiredCache,
-  isComplexSearch,
-  type SearchCache,
-} from '@/lib/utils/taskFiltering';
 import { validateTaskCollection } from '@/lib/utils/taskValidation';
 import { logger } from '@/lib/utils/logger';
 
@@ -233,7 +224,7 @@ export function createApplyFilters(get: () => TaskStoreState, set: StoreSetter) 
       // Validate board access for cross-board search
       if (filters.crossBoardSearch) {
         const beforeCount = validTasks.length;
-        validTasks = await validateBoardAccess(validTasks);
+        validTasks = await filterAccessibleTasks(validTasks);
         if (validTasks.length !== beforeCount) tasksChanged = true;
       }
 
@@ -386,4 +377,154 @@ export function createSaveSearchScope() {
       logger.warn('Failed to save search scope preference:', error);
     }
   };
+}
+
+// ---------------------------------------------------------------------------
+// Search-result cache + filtering helpers
+//
+// These live here, beside the `searchCache` state they operate on and the
+// `applyFilters` action that drives them, rather than in a separate utility
+// module the store had to pass its own state into. taskSearch.ts stays a pure,
+// reusable text-search module; only the cache lifecycle is co-located here.
+// ---------------------------------------------------------------------------
+
+const CACHE_TTL = 5 * 60 * 1000; // 5 minutes
+const MAX_CACHE_SIZE = 50;
+
+type CacheEntry = { results: Task[]; timestamp: number };
+export type SearchCache = Map<string, CacheEntry>;
+
+/**
+ * Filters a task list down to the ones whose board still exists and is not
+ * archived — used by cross-board search to drop tasks from gone boards.
+ */
+export async function filterAccessibleTasks(tasks: Task[]): Promise<Task[]> {
+  try {
+    const boards = await taskDB.getBoards();
+    const validBoardIds = new Set<string>();
+    for (const b of boards) {
+      if (!b.archivedAt) validBoardIds.add(b.id);
+    }
+
+    const accessibleTasks = tasks.filter(task => validBoardIds.has(task.boardId));
+    if (accessibleTasks.length !== tasks.length) {
+      logger.info('Filtered tasks from inaccessible boards', {
+        filteredCount: tasks.length - accessibleTasks.length,
+      });
+    }
+    return accessibleTasks;
+  } catch (error: unknown) {
+    logger.warn('Failed to validate board access, proceeding with existing tasks', error);
+    return tasks;
+  }
+}
+
+/**
+ * Generates cache key for search filters
+ */
+export function generateCacheKey(filters: TaskFilters): string {
+  const key = {
+    search: filters.search,
+    status: filters.status,
+    priority: filters.priority,
+    tags: filters.tags.toSorted(),
+    boardId: filters.boardId,
+    crossBoardSearch: filters.crossBoardSearch,
+    dateRange: filters.dateRange ? {
+      start: filters.dateRange.start.getTime(),
+      end: filters.dateRange.end.getTime()
+    } : null
+  };
+  return JSON.stringify(key);
+}
+
+/**
+ * Checks cache for existing search results
+ * Returns cached results if valid, null otherwise
+ */
+export function checkCache(
+  cacheKey: string,
+  searchCache: SearchCache,
+  currentTasks: Task[]
+): Task[] | null {
+  if (!searchCache.has(cacheKey)) return null;
+
+  const cached = searchCache.get(cacheKey);
+  if (!cached) return null;
+  const now = Date.now();
+
+  // Check if cache is expired
+  if (now - cached.timestamp >= CACHE_TTL) {
+    searchCache.delete(cacheKey);
+    return null;
+  }
+
+  // Validate cached results still exist
+  const currentTaskIds = new Set(currentTasks.map(t => t.id));
+  const validCachedResults = cached.results.filter(task => currentTaskIds.has(task.id));
+
+  if (validCachedResults.length === cached.results.length) {
+    return validCachedResults;
+  }
+
+  // Cache is stale
+  searchCache.delete(cacheKey);
+  return null;
+}
+
+/**
+ * Caches search results
+ * Manages cache size and cleans up old entries
+ */
+export function cacheResults(
+  cacheKey: string,
+  results: Task[],
+  searchCache: SearchCache
+): void {
+  try {
+    // Clean up old cache entries if at max size
+    if (searchCache.size >= MAX_CACHE_SIZE) {
+      const entriesToRemove = Math.floor(MAX_CACHE_SIZE * 0.2); // Remove 20% of entries
+      const keys = Array.from(searchCache.keys());
+      for (let i = 0; i < entriesToRemove; i++) {
+        searchCache.delete(keys[i]);
+      }
+    }
+
+    searchCache.set(cacheKey, {
+      results,
+      timestamp: Date.now()
+    });
+  } catch (error: unknown) {
+    logger.warn('Failed to cache search results', error);
+    searchCache.clear();
+  }
+}
+
+/**
+ * Cleans up expired cache entries periodically
+ */
+export function cleanupExpiredCache(searchCache: SearchCache): void {
+  try {
+    const now = Date.now();
+    for (const [key, value] of searchCache.entries()) {
+      if (now - value.timestamp > CACHE_TTL) {
+        searchCache.delete(key);
+      }
+    }
+  } catch (error: unknown) {
+    logger.warn('Cache cleanup failed', error);
+  }
+}
+
+/**
+ * Determines if search operation is complex
+ */
+export function isComplexSearch(tasks: Task[], filters: TaskFilters): boolean {
+  return !!(filters.search && (
+    tasks.length > 200 ||
+    filters.crossBoardSearch ||
+    (filters.tags && filters.tags.length > 0) ||
+    !!filters.dateRange
+  ));
 }
